backend/requirements.txt
flask==3.0.0
streamlit==1.32.0
groq==0.5.0
pdfplumber==0.10.3
python-dotenv==1.0.1
requests==2.31.0